feat(events.py): add shutdown and shutup commands to bot
Adds two new commands to the bot: `shutdown` and `shutup`. The `shutdown` command shuts down the bot if the user invoking the command is the owner of the bot. The `shutup` command sets the bot's status to invisible if the user invoking the command is the owner of the bot.